requirements.txt
Django-Gtts
phonenumbers
django_twilio
facebook-sdk
tweepy
smtp-mail
piplapis-python
